Move TxMeta to xrpl_core
* add peekNodes() to TxMeta for const access to nodes
