+++ src/ripple/app/ledger/Ledger.h
@@ -24,12 +24,12 @@
 #include <ripple/beast/utility/Journal.h>
 #include <ripple/core/TimeKeeper.h>
 #include <ripple/ledger/CachedView.h>
-#include <ripple/ledger/TxMeta.h>
 #include <ripple/ledger/View.h>
 #include <ripple/protocol/Book.h>
 #include <ripple/protocol/Indexes.h>
 #include <ripple/protocol/STLedgerEntry.h>
 #include <ripple/protocol/Serializer.h>
+#include <ripple/protocol/TxMeta.h>
 #include <ripple/shamap/SHAMap.h>
 #include <mutex>
