+++ src/ripple/app/misc/Transaction.h
@@ -22,11 +22,11 @@
 
 #include <ripple/basics/RangeSet.h>
 #include <ripple/beast/utility/Journal.h>
-#include <ripple/ledger/TxMeta.h>
 #include <ripple/protocol/ErrorCodes.h>
 #include <ripple/protocol/Protocol.h>
 #include <ripple/protocol/STTx.h>
 #include <ripple/protocol/TER.h>
+#include <ripple/protocol/TxMeta.h>
 #include <boost/optional.hpp>
 #include <optional>
 #include <variant>
