+++ src/ripple/ledger/detail/ApplyStateTable.h
@@ -25,8 +25,8 @@
 #include <ripple/ledger/OpenView.h>
 #include <ripple/ledger/RawView.h>
 #include <ripple/ledger/ReadView.h>
-#include <ripple/ledger/TxMeta.h>
 #include <ripple/protocol/TER.h>
+#include <ripple/protocol/TxMeta.h>
 #include <memory>
 
 namespace ripple {
